+++ src/main/python/apache/thermos/common/process_util.py
@@ -12,7 +12,10 @@
 # limitations under the License.
 #
 
+import ctypes
 import os
+
+from twitter.common import log
 
 from gen.apache.aurora.api.constants import TASK_FILESYSTEM_MOUNT_POINT
 
@@ -42,3 +45,30 @@ def wrap_with_mesos_containerizer(cmdline, user, cwd, mesos_containerizer_path):
               os.path.join(os.environ['MESOS_DIRECTORY'], TASK_FILESYSTEM_MOUNT_POINT),
               user,
               bash_wrapper % cmdline))
+
+
+def setup_child_subreaping():
+  """
+  This uses the prctl(2) syscall to set the `PR_SET_CHILD_SUBREAPER` flag. This
+  means if any children processes need to be reparented, they will be reparented
+  to this process.
+
+  More documentation here: http://man7.org/linux/man-pages/man2/prctl.2.html
+  and here: https://lwn.net/Articles/474787/
+
+  Callers should reap terminal children to prevent zombies.
+
+  raises OSError if the underlying prctl call fails.
+  raises RuntimeError if libc cannot be found.
+  """
+  log.debug("Calling prctl(2) with PR_SET_CHILD_SUBREAPER")
+  # This constant is taken from prctl.h
+  PR_SET_CHILD_SUBREAPER = 36
+  library_name = ctypes.util.find_library('c')
+  if library_name is None:
+    raise RuntimeError("libc not found")
+  libc = ctypes.CDLL(library_name, use_errno=True)
+  ret = libc.prctl(PR_SET_CHILD_SUBREAPER, 1, 0, 0, 0)
+  if ret != 0:
+    errno = ctypes.get_errno()
+    raise OSError(errno, os.strerror(errno))